requirements.txt
<<<<<<< HEAD
Django>=4.2
djangorestframework
djangorestframework-simplejwt
=======
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
sqlparse==0.5.3
>>>>>>> 7ef53d3 (Initial working Django REST API (clients CRUD))
